Updated imageProcessing.py
Updated code, originally I copy-pasted the same code for different images, and changed the code slightly for each different image. In this update I changed it so that I could simply pass it through a function called imageProcessingFunction() which does the same thing the previous code did, except I don't have to copy-paste the code and change it slightly for each image. All required the details for the function are given in the header.